fix: upload and update logic for talent and employer

+++ src/controllers/employerProfile.controller.ts
@@ -10,38 +10,59 @@ import type { Request, Response } from "express";
 import { StatusCodes } from "http-status-codes";
 
 class EmployerProfileController {
-  public createEmployerProfile = catchAsync(async (req: Request, res: Response) => {
+  public createOrUpdateEmployerProfile = catchAsync(async (req: Request, res: Response) => {
     if (!req.user) {
       throw new ApiError(StatusCodes.NOT_FOUND, "User not found");
     }
-    const files = req.files as { [fieldname: string]: Express.Multer.File[] };
 
-    const bannerFile = files.banner[0];
-    const logoFile = files.logo[0];
-    const bannerFilePath = path.join(bannerFile.destination, bannerFile.filename);
-    const logoFilePath = path.join(logoFile.destination, logoFile.filename);
+    let bannerUrl = "";
+    let logoUrl = "";
 
-    console.log(logoFilePath, bannerFilePath);
-
-    let bannerUrl: string;
-    let logoUrl: string;
-    try {
-      bannerUrl = await handleImageUpload(bannerFile.filename, bannerFilePath, BucketNameEnum.BANNER);
-      logoUrl = await handleImageUpload(logoFile.filename, logoFilePath, BucketNameEnum.LOGO);
-    } catch (error) {
-      if (fs.existsSync(logoFilePath)) {
-        fs.unlinkSync(logoFilePath);
-      }
-      if (fs.existsSync(bannerFilePath)) {
-        fs.unlinkSync(bannerFilePath);
-      }
-      throw new ApiError(
-        StatusCodes.INTERNAL_SERVER_ERROR,
-        `Failed·to·upload·logo·or·banner·image: ${(error as Error).message}`,
-      );
-    }
     const userId = req.user.id;
-    const serviceResponse = await employerProfileService.createEmployerProfile(req.body, bannerUrl, logoUrl, userId);
+    if (req.files) {
+      const files = req.files as { [fieldname: string]: Express.Multer.File[] };
+      if (files?.banner?.[0]) {
+        const bannerFile = files.banner[0];
+        const bannerFilePath = path.join(bannerFile.destination, bannerFile.filename);
+
+        try {
+          bannerUrl = await handleImageUpload(bannerFile.filename, userId, bannerFilePath, BucketNameEnum.BANNER);
+        } catch (error) {
+          if (fs.existsSync(bannerFilePath)) {
+            fs.unlinkSync(bannerFilePath);
+          }
+          throw new ApiError(
+            StatusCodes.INTERNAL_SERVER_ERROR,
+            `Failed to upload banner image: ${(error as Error).message}`,
+          );
+        }
+      }
+
+      if (files?.logo?.[0]) {
+        const logoFile = files.logo[0];
+        const logoFilePath = path.join(logoFile.destination, logoFile.filename);
+
+        try {
+          logoUrl = await handleImageUpload(logoFile.filename, userId, logoFilePath, BucketNameEnum.LOGO);
+        } catch (error) {
+          if (fs.existsSync(logoFilePath)) {
+            fs.unlinkSync(logoFilePath);
+          }
+          throw new ApiError(
+            StatusCodes.INTERNAL_SERVER_ERROR,
+            `Failed to upload logo image: ${(error as Error).message}`,
+          );
+        }
+      }
+    }
+
+    const serviceResponse = await employerProfileService.createOrUpdateEmployerProfile(
+      req.body,
+      bannerUrl,
+      logoUrl,
+      userId,
+    );
+
     return handleServiceResponse(serviceResponse, res);
   });
 
@@ -56,15 +77,19 @@ class EmployerProfileController {
     return handleServiceResponse(serviceResponse, res);
   });
 
-  public updateEmployerProfile = catchAsync(async (req: Request, res: Response) => {
-    const { id } = req.params;
-    const serviceResponse = await employerProfileService.updateEmployerProfile(id, req.body);
-    return handleServiceResponse(serviceResponse, res);
-  });
-
   public deleteEmployerProfile = catchAsync(async (req: Request, res: Response) => {
     const { id } = req.params;
     const serviceResponse = await employerProfileService.deleteEmployerProfile(id);
+    return handleServiceResponse(serviceResponse, res);
+  });
+
+  public getRegisteredEmployerProfile = catchAsync(async (req: Request, res: Response) => {
+    if (!req.user) {
+      throw new ApiError(StatusCodes.NOT_FOUND, "User not found");
+    }
+
+    const userId = req.user.id;
+    const serviceResponse = await employerProfileService.getRegisteredEmployerProfile(userId);
     return handleServiceResponse(serviceResponse, res);
   });
 }
